% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{freq_fx}
\alias{freq_fx}
\title{frequency}
\usage{
freq_fx(vec, na.rm = TRUE, digits = 1)
}
\arguments{
\item{vec}{vector in which have only two values, 0 and 1.}

\item{na.rm}{na.rm}

\item{digits}{digits}
}
\value{
percent
}
\description{
rounds the values in its first argument to the specified number of decimal places (default 0).
}
