% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making table.R
\name{Mean_CI}
\alias{Mean_CI}
\title{insert mean and confidence interval into table}
\usage{
Mean_CI(x, digits = 3)
}
\arguments{
\item{x}{a vector which has mean ans confidence interval.}

\item{digits}{integer indicating the number of (fixed) decimal places.}
}
\value{
a character atomic vector
}
\description{
insert mean and confidence interval into table
}
