% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making table.R
\name{oddsratio_CI}
\alias{oddsratio_CI}
\title{insert Odds ratio into table}
\usage{
oddsratio_CI(coef, conf, digits)
}
\arguments{
\item{coef}{a atomic vector which has a coefficient value.}

\item{conf}{a vector which has low limit and upper limir of confidence interval.}

\item{digits}{integer indicating the number of (fixed) decimal places.}
}
\value{
a character atomic vector
}
\description{
insert Odds ratio into table
}
