% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_character.R
\name{order_character}
\alias{order_character}
\title{order_character}
\usage{
order_character(vec_cha, vec_or)
}
\arguments{
\item{vec_cha}{atomic character vector.}

\item{vec_or}{atomic character vector having a specific order.}
}
\value{
order vector
}
\description{
order_character returns a permutation which rearrange its first argument into specified order of its second argument.
}
