% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_new.R
\name{round_new}
\alias{round_new}
\title{Rounding Numberss}
\usage{
round_new(x, digits = 3)
}
\arguments{
\item{x}{A numeric vector to be rounded.}

\item{digits}{integer indicating the number of (fixed) decimal places.}
}
\value{
A numeric vectors
}
\description{
rounds the values in its first argument to the specified number of decimal places (default 0).
}
