% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underbar_change.R
\name{underbar_change}
\alias{underbar_change}
\title{change to underbar}
\usage{
underbar_change(x)
}
\arguments{
\item{x}{the vector have marks or symbols changed to underbar.}
}
\value{
A vector
}
\description{
underbar_change
}
